chore: start introducing forward declaration headers and split app functionality in application lib
