+++ libs/application/src/update.cpp
@@ -1,22 +1,19 @@
 //
-// lager - library for functional interactive c++ programs
-// Copyright (C) 2017 Juan Pedro Bolivar Puente
-//
-// This file is part of lager.
-//
-// lager is free software: you can redistribute it and/or modify
-// it under the terms of the MIT License, as detailed in the LICENSE
-// file located at the root of this source code distribution,
-// or here: <https://github.com/arximboldi/lager/blob/master/LICENSE>
+// Created by nickel on 11/30/24.
 //
 
+#include <bulin/application/update.hpp>
 #include <bulin/application/app.hpp>
 
 #include <bulin/graphics/shader_data.hpp>
 #include <bulin/graphics/shader_model.hpp>
 #include <bulin/graphics/texture.hpp>
 
+#include <lager/extra/struct.hpp>
+
 #include <iostream>
+
+LAGER_STRUCT(bulin, app, doc, path);
 
 namespace bulin
 {
@@ -25,7 +22,7 @@ app_result update(app app_state, app_action app_action)
   return lager::match(std::move(app_action))(
       [&](save_action&& save_action) -> app_result
       {
-        app_state.path = save_action.file.replace_extension("bulin");
+        app_state.path = std::move(save_action).file.replace_extension("bulin");
         auto eff = [model = app_state.doc, filepath = app_state.path](auto&&)
         {
           std::cout << "saving file: " << filepath << '\n';
@@ -59,5 +56,4 @@ app_result update(app app_state, app_action app_action)
         return {std::move(app_state), eff};
       });
 }
-
-}  // namespace bulin
+}
