+++ libs/graphics/CMakeLists.txt
@@ -20,6 +20,7 @@ target_sources(bulin_graphics
     ${CMAKE_CURRENT_BINARY_DIR}
   FILES
     include/bulin/graphics/flat_shader.hpp
+    include/bulin/graphics/graphics_fwd.hpp
     include/bulin/graphics/shader_data.hpp
     include/bulin/graphics/shader_model.hpp
     include/bulin/graphics/texture.hpp
