+++ libs/graphics/include/bulin/graphics/graphics_fwd.hpp
@@ -1,0 +1,12 @@
+//
+// Created by nickel on 11/30/24.
+//
+
+#pragma once
+
+namespace bulin
+{
+struct shader_data;
+class shader_model;
+class texture;
+}
